Correciones

+++ app/src/main/java/co/edu/unab/castellanos/jose/restaurantfastfood_admin/model/entity/Product.java
@@ -10,7 +10,7 @@ public class Product implements Serializable {
     private String id;
     private String type;
     private String name;
-    private double price;
+    private int price;
     private String description;
     private String url_picture;
 
@@ -18,7 +18,7 @@ public class Product implements Serializable {
 
     }
 
-    public Product(String type,String name, double price, String description, String url_picture) {
+    public Product(String type,String name, int price, String description, String url_picture) {
 
         this.type = type;
         this.name = name;
@@ -67,7 +67,7 @@ public class Product implements Serializable {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
